Manually set offsetTop and offsetWidth for svgArea.
This is to work-around what I think is a bug in Firefox.  (Known bug,
there seem to be a few reports of offsetWidth/etc bugs in Firefox, but
many marked unconfirmed or wontfix, e.g. 458617.  However, this
problem persists in the current nightly build, so working around by
setting the property manually.)