Add Load/Save buttons, implement basics of saving.
Adds a row at the top containing links called "Load" and "Save".
"Save" generates a JSON file, its href is updated dynamically so that
it points to a link that contains the device name as an argument.  The
user can click it and save with the file name generated by the server
(basically "<devicename>.json") or can right-click and Save Link As to
choose a different file name.

Loading still to do, as well as formatting file contents correctly per
the JSON file format used by the Max/MSP GUI.