Implement file loading using the hidden iframe trick for background uploading.
Adds a file selection form to the display when "Load" is clicked.
When user clicks it and chooses a file, the form is submitted with the
target being a temporary hidden iframe.  The POST is handled by the
Python server, interpreting the JSON and sending /link, /connect, or
/connection/modify messages as specified in the file.

In the future, we could use HTML5 local file handling to make this
slightly smoother, such as automatically opening the file selection
dialog, and allowing drag & drop.