Automatically close the server when all connections are lost for a time.
While this is not guaranteed to be 100% solid, it seems to work.  For
polling method, it should be fairly reliable.  For websockets, there
is a 10-second window for it to open a new websocket connection, and
it will try every second to do so.

Detecting that no connections are available should allow to end the
program after user closes the app window. There is no way to be sure
the browser app window will send any notification, so this method is
prefered.